♻️Refactor: useObserver 분리
